removed _QCircuitOperation
